Added a WireVector indexer based on NameIndexer, replaced locations that used next_tempvar_name, etc. with methods of this class
